+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## 3.5.0 -- UNRELEASED
 
 - Default style for `print-table` can be overridden via a dynamic var
+- Use a vertical bar (│) not a pipe character (|) in binary output
 
 
 ## 3.4.1 -- 23 Jun 2025